This fixes not being able to edit tasks with notes attached. (#58)
When the feature of adding notes to tasks was added, it put an emoji on
the text line. The regex used to match tasks was failing because it
wasn't account for the emoji.

This change removes the start of string check in the regex to keep it
back to working.